django-rest-api-app/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
django-cors-headers
python-dotenv
Pillow
pypdf
PyMuPDF